# backend/src/swagger.yaml
openapi: "3.0.3"
info:
  title: HalalChain V2 API
  description: Enterprise Halal Supply Chain Platform — REST API
  version: "2.0.0"
  contact:
    name: HalalChain Team
servers:
  - url: http://localhost:4000
    description: Local development server
  - url: https://api.halalchain.app
    description: Production server

# ─────────────────────────────────────────────────────────────
# Security
# ─────────────────────────────────────────────────────────────
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token

  # ─────────────────────────────────────────────────────────────
  # Schemas
  # ─────────────────────────────────────────────────────────────
  schemas:
...
    Supplier:
      type: object
      required:
        - id
        - name
        - country
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440001"
        name:
          type: string
          example: "Al-Barakah Foods Sdn Bhd"
        country:
          type: string
          example: "Malaysia"
        email:
          type: string
          format: email
          nullable: true
          example: "contact@albarakah.com"
        phone:
          type: string
          nullable: true
          example: "+60123456789"
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
          example: "ACTIVE"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-10T06:00:00.000Z"

    # ── Product ────────────────────────────────────────────────
    Product:
      type: object
      required:
        - id
        - supplierId
        - name
        - sku
        - category
        - unit
        - unitPrice
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: "770e8400-e29b-41d4-a716-446655440002"
        supplierId:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440001"
        name:
          type: string
          example: "Halal Chicken Breast"
        sku:
          type: string
          example: "HCB-001"
        category:
          type: string
          example: "Poultry"
        unit:
          type: string
          example: "kg"
        unitPrice:
          type: number
          format: float
          example: 12.50
        createdAt:
          type: string
          format: date-time
          example: "2024-01-12T09:00:00.000Z"
        supplier:
          $ref: "#/components/schemas/Supplier"
        qrCodeUrl:
          type: string
          description: Base64-encoded PNG data URL for the product QR code
          example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."

    # ── HalalCertificate ───────────────────────────────────────
...
    PaginatedResponse:
      type: object
      required:
        - data
        - page
        - limit
        - total
        - totalPages
      properties:
        data:
          type: array
          items:
            type: object
          description: Array of result items for the current page
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          minimum: 0
          example: 85
        totalPages:
          type: integer
          minimum: 0
          example: 5

    # ── ErrorResponse ──────────────────────────────────────────
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Resource not found"

# ─────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────
paths:
...
  /api/products:
    get:
      tags:
        - Products
      summary: List products (paginated, filterable)
      description: Returns a paginated list of products. Supports filtering by name and SKU. Requires ADMIN or MANAGER role.
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: name
          in: query
          description: Filter by product name (case-insensitive, partial match)
          schema:
            type: string
            example: "Chicken"
        - name: sku
          in: query
          description: Filter by SKU (case-insensitive, partial match)
          schema:
            type: string
            example: "HCB"
      responses:
        "200":
          description: Paginated list of products
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Product"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags:
        - Products
      summary: Create a new product (ADMIN only)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - supplierId
                - name
                - sku
                - category
                - unit
              properties:
                supplierId:
                  type: string
                  format: uuid
                  example: "660e8400-e29b-41d4-a716-446655440001"
                name:
                  type: string
                  minLength: 2
                  maxLength: 200
                  example: "Halal Chicken Breast"
                sku:
                  type: string
                  minLength: 2
                  maxLength: 50
                  example: "HCB-001"
                category:
                  type: string
                  minLength: 2
                  maxLength: 100
                  example: "Poultry"
                unit:
                  type: string
                  minLength: 1
                  maxLength: 30
                  example: "kg"
                unitPrice:
                  type: number
                  minimum: 0
                  example: 12.50
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                type: object
                properties:
                  product:
                    $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Supplier not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: SKU already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "SKU already exists"
  /api/products/{id}:
    get:
      tags:
        - Products
      summary: Get a single product by ID (includes QR code URL)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Product details including qrCodeUrl
          content:
            application/json:
              schema:
                type: object
                properties:
                  product:
                    $ref: "#/components/schemas/Product"
                  qrCodeUrl:
                    type: string
                    description: Base64-encoded PNG data URL of the product QR code
                    example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      tags:
        - Products
      summary: Update a product (ADMIN only)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 200
                sku:
                  type: string
                  minLength: 2
                  maxLength: 50
                category:
                  type: string
                  minLength: 2
                  maxLength: 100
                unit:
                  type: string
                  minLength: 1
                  maxLength: 30
                unitPrice:
                  type: number
                  minimum: 0
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  product:
                    $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - Products
      summary: Delete a product (ADMIN only)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Product deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Product deleted"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/products/{id}/traceability:
    get:
      tags:
        - Products
      summary: Get full supply chain traceability timeline for a product
      description: Returns the product details and a chronological timeline of all supply chain events — supplier, halal certificates, purchase orders, shipments, and inventory movements. Requires ADMIN or MANAGER role.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the product
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Traceability timeline with all supply chain events
          content:
            application/json:
              schema:
                type: object
                required:
                  - product
                  - timeline
                properties:
                  product:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      name:
                        type: string
                      sku:
                        type: string
                  timeline:
                    type: array
                    items:
                      type: object
                      required:
                        - type
                        - date
                        - title
                        - description
                        - metadata
                      properties:
                        type:
                          type: string
                          enum:
                            - SUPPLIER
                            - CERTIFICATE
                            - PRODUCT
                            - PURCHASE_ORDER
                            - SHIPMENT
                            - INVENTORY
                        date:
                          type: string
                          format: date-time
                        title:
                          type: string
                        description:
                          type: string
                        metadata:
                          type: object
                          additionalProperties: true
              example:
                product:
                  id: "770e8400-e29b-41d4-a716-446655440002"
                  name: "Halal Chicken Breast"
                  sku: "HCB-001"
                timeline:
                  - type: "SUPPLIER"
                    date: "2024-01-10T06:00:00.000Z"
                    title: "Supplier: Al-Barakah Foods Sdn Bhd"
                    description: "Malaysia · ACTIVE"
                    metadata:
                      supplierId: "660e8400-e29b-41d4-a716-446655440001"
                      supplierName: "Al-Barakah Foods Sdn Bhd"
                      country: "Malaysia"
                      status: "ACTIVE"
                  - type: "CERTIFICATE"
                    date: "2024-01-01T00:00:00.000Z"
                    title: "Certificate: JAKIM/HPC/2024/001234"
                    description: "Issued by JAKIM · Expires 2025-12-31"
                    metadata:
                      certificateId: "880e8400-e29b-41d4-a716-446655440003"
                      certificateNumber: "JAKIM/HPC/2024/001234"
                      issuedDate: "2024-01-01T00:00:00.000Z"
                      expiryDate: "2025-12-31T00:00:00.000Z"
                      issuingBody: "JAKIM"
                  - type: "PRODUCT"
                    date: "2024-01-12T09:00:00.000Z"
                    title: "Product: Halal Chicken Breast"
                    description: "SKU: HCB-001"
                    metadata:
                      productId: "770e8400-e29b-41d4-a716-446655440002"
                      name: "Halal Chicken Breast"
                      sku: "HCB-001"
                  - type: "PURCHASE_ORDER"
                    date: "2024-05-20T11:00:00.000Z"
                    title: "Purchase Order: PO-2024-00042"
                    description: "Status: APPROVED · Total: 3750.00"
                    metadata:
                      poId: "cc0e8400-e29b-41d4-a716-446655440007"
                      poNumber: "PO-2024-00042"
                      status: "APPROVED"
                      totalAmount: "3750.00"
                  - type: "SHIPMENT"
                    date: "2024-06-02T06:00:00.000Z"
                    title: "Shipment: TRK-2024-98765"
                    description: "Kuala Lumpur, Malaysia → Jakarta, Indonesia · IN_TRANSIT"
                    metadata:
                      shipmentId: "ee0e8400-e29b-41d4-a716-446655440009"
                      trackingNumber: "TRK-2024-98765"
                      status: "IN_TRANSIT"
                      origin: "Kuala Lumpur, Malaysia"
                      destination: "Jakarta, Indonesia"
                  - type: "INVENTORY"
                    date: "2024-06-05T08:30:00.000Z"
                    title: "Inventory INBOUND: 300 units"
                    description: "Warehouse: Warehouse A"
                    metadata:
                      movementId: "bb0e8400-e29b-41d4-a716-446655440006"
                      type: "INBOUND"
                      quantity: 300
                      warehouseName: "Warehouse A"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role — ADMIN or MANAGER required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Product not found"

  # ══════════════════════════════════════════════════════════
  # Certificates
  # ══════════════════════════════════════════════════════════
...
  /api/public/products/{id}/traceability:
    get:
      tags:
        - Public
      summary: Get public traceability data for a product (no auth required)
      description: Public endpoint — no JWT cookie required. Returns product info, supplier details, halal certificates, shipments, and the traceability timeline. Used by the consumer-facing QR scan page.
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the product
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Public traceability data
          content:
            application/json:
              schema:
                type: object
                required:
                  - product
                  - supplier
                  - certificates
                  - shipments
                  - timeline
                properties:
                  product:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      name:
                        type: string
                      sku:
                        type: string
                  supplier:
                    type: object
                    properties:
                      name:
                        type: string
                      country:
                        type: string
                  certificates:
                    type: array
                    items:
                      type: object
                      properties:
                        certificateNumber:
                          type: string
                        expiryDate:
                          type: string
                          format: date
                        issuedBy:
                          type: string
                  shipments:
                    type: array
                    items:
                      type: object
                      properties:
                        trackingNumber:
                          type: string
                        status:
                          type: string
                  timeline:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                        date:
                          type: string
                          format: date-time
                        title:
                          type: string
                        description:
                          type: string
                        metadata:
                          type: object
                          additionalProperties: true
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Product not found"